Fixed the DynamicHelper class to map the Integer DynamicType to the Long CLR type

+++ src/Serialization/Neuroglia.Serialization.Dynamic/DynamicHelper.cs
@@ -100,7 +100,7 @@ namespace Neuroglia.Serialization
                 DynamicType.Boolean => typeof(bool),
                 DynamicType.Timestamp => typeof(Timestamp),
                 DynamicType.Duration => typeof(Duration),
-                DynamicType.Integer => typeof(int),
+                DynamicType.Integer => typeof(long),
                 DynamicType.Number => typeof(double),
                 DynamicType.Array => typeof(DynamicArray),
                 DynamicType.Object => typeof(DynamicObject),
